+++ src/modules/growl/growl.service.ts
@@ -19,7 +19,7 @@ import {
 import {GroundwaterLevelsService} from "../../api/groundwater-levels.service";
 import {GeoDataService} from "../../api/geo-data.service";
 import nlwknMeasurementClassificationColors from "../../assets/nlwkn-measurement-classification-colors.toml";
-import {utils} from "../../common/utils";
+import {range} from "../../common/utils/range";
 
 export namespace GrowlService {
   // geo data
@@ -203,7 +203,7 @@ export class GrowlService {
     }
 
     let today = dayjs();
-    let week = utils.range(7);
+    let week = range(7);
     let measurements = Object.fromEntries(
       week.map(i => {
         let day = today.subtract(dayjs.duration(i, "day"));
